# backend/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A social network API built with Go. Supports user authentication, posts, comments, reactions, groups, messaging, notifications, and real-time WebSocket events.",
        "title": "Social Network API",
        "termsOfService": "https://example.com/terms",
        "contact": {
            "name": "API Support",
            "url": "https://example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/comments/create": {
            "post": {
                "description": "Adds a comment to a post. Accepts JSON or multipart/form-data with optional image.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comments"
                ],
                "summary": "Create a comment",
                "parameters": [
                    {
                        "description": "Post ID (for JSON requests)",
                        "name": "postId",
                        "in": "body",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Comment text",
                        "name": "text",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Post ID (for multipart requests)",
                        "name": "postId",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Comment body",
                        "name": "text",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Optional comment image",
                        "name": "image",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Comment created successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Not logged in",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/comments/{id}/delete": {
            "delete": {
                "description": "Handles like/dislike on comments and comment deletion.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comments"
                ],
                "summary": "Comment operations",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Comment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/comments/{id}/dislike": {
            "post": {
                "description": "Handles like/dislike on comments and comment deletion.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comments"
                ],
                "summary": "Comment operations",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Comment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/comments/{id}/like": {
            "post": {
                "description": "Handles like/dislike on comments and comment deletion.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comments"
                ],
                "summary": "Comment operations",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Comment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/follow/{resolver}/{id}": {
            "put": {
                "description": "Follow, unfollow, accept follow request, or reject follow request.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Follow"
                ],
                "summary": "Follow operations",
                "parameters": [
                    {
                        "enum": [
                            "follow",
                            "unfollow",
                            "accept",
                            "reject"
                        ],
                        "type": "string",
                        "description": "Action: follow, unfollow, accept, or reject",
                        "name": "resolver",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Target user ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Action successful",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Not logged in",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "SessionCookie": {
            "type": "apiKey",
            "name": "session_id",
            "in": "cookie"
        }
    }
}